+++ src/Tests/Aardvark.Rendering.Tests/Tests/Other/SgTest.fs
@@ -7,7 +7,6 @@ open Aardvark.Base
 open Aardvark.Rendering
 open Aardvark.SceneGraph
 open Aardvark.SceneGraph.Semantics
-open Aardvark.Application
 open FSharp.Data.Adaptive
 open Expecto
 open FShade
@@ -44,7 +43,7 @@ module ``SceneGraph Tests`` =
 
         let rec prettyName (t : Type) =
             match intrisicNames.TryGetValue t with
-                | (true, n) -> n
+                | true, n -> n
                 | _ ->
                     if t.IsArray then
                         sprintf "%s[]" (t.GetElementType() |> prettyName)
@@ -61,7 +60,7 @@ module ``SceneGraph Tests`` =
 
             let sgModule = typeof<Sg.Set>.DeclaringType
 
-            for (att, expected) in semantics do
+            for att, expected in semantics do
                 for t in sgTypes do
                     if t.DeclaringType = sgModule then
                         let hasRule = Ag.hasSynRule t expected att
@@ -166,10 +165,262 @@ module ``SceneGraph Tests`` =
             app.Complete()
         }
 
+    let modelTrafo =
+        test "Sg.ModelTrafo" {
+            IntrospectionProperties.CustomEntryAssembly <- Assembly.GetAssembly(typeof<ISg>)
+            Aardvark.Init()
+
+            let translation = V3d(40.0, 12.0, -23.0)
+            let rotation = V3d(3.0, -2.0, 1.0)
+            let scaling = V3d(0.5, 2.0, 3.0)
+
+            let sg =
+                Sg.fullScreenQuad
+                |> Sg.translation' translation
+                |> Sg.rotation' rotation
+                |> Sg.scaling' scaling
+
+            let ro =
+                sg.RenderObjects(Ag.Scope.Root).Content.GetValue()
+                |> Seq.head
+
+            let result = ro.AttributeScope.ModelTrafo |> AVal.force
+            let expected = Trafo3d.Translation translation * Trafo3d.RotationEuler rotation * Trafo3d.Scale scaling
+
+            Expect.approxEquals result.Forward expected.Forward 0.001 "Invalid model trafo"
+        }
+
+    module BoundingBox =
+
+        let private randomTrafo() =
+            let s = (Rnd.v3d() + 0.1) * 2.0
+            let t = (Rnd.v3d() - 0.5) * 10.0
+            let r = (Rnd.v3d() - 0.5) * Constant.PiTimesFour
+            Trafo3d.Scale s * Trafo3d.Translation t * Trafo3d.RotationEuler r
+
+        let private drawLeaf (trafo: aval<Trafo3d>) (positions: V3f[]) =
+            Sg.draw IndexedGeometryMode.TriangleList
+            |> Sg.vertexArray DefaultSemantic.Positions positions
+            |> Sg.trafo trafo
+
+        let renderNode =
+            test "Bounding Box.RenderNode" {
+                IntrospectionProperties.CustomEntryAssembly <- Assembly.GetAssembly(typeof<ISg>)
+                Aardvark.Init()
+
+                let positions = [| V3f(-0.5f, -0.25f, 0.0f); V3f(0.0f, -10.0f, 5.0f); V3f(3.0f, 7.0f, -7.0f) |]
+
+                let translation = V3d(40.0, 12.0, -23.0)
+                let rotation = V3d(3.0, -2.0, 1.0)
+                let scaling = V3d(0.5, 2.0, 3.0)
+                let trafo = Trafo3d.Translation translation * Trafo3d.RotationEuler rotation * Trafo3d.Scale scaling
+
+                let sg =
+                    Sg.draw IndexedGeometryMode.TriangleList
+                    |> Sg.vertexArray DefaultSemantic.Positions positions
+                    |> Sg.translation' translation
+                    |> Sg.rotation' rotation
+                    |> Sg.scaling' scaling
+
+                let expected = positions |> Array.map (V3d >> Mat.transformPos trafo.Forward) |> Box3d
+                let lbb = sg.LocalBoundingBox Ag.Scope.Root |> AVal.force
+                let gbb = sg.GlobalBoundingBox Ag.Scope.Root |> AVal.force
+
+                Expect.approxEquals lbb expected 0.001 "Invalid local bounding box"
+                Expect.approxEquals gbb expected 0.001 "Invalid global bounding box"
+            }
+
+        let renderObjectsNode =
+            test "Bounding Box.RenderObjectsNode" {
+                IntrospectionProperties.CustomEntryAssembly <- Assembly.GetAssembly(typeof<ISg>)
+                Aardvark.Init()
+
+                let positions = [| V3f(-0.5f, -0.25f, 0.0f); V3f(0.0f, -10.0f, 5.0f); V3f(3.0f, 7.0f, -7.0f) |]
+                let globalTrafo = AVal.init <| randomTrafo()
+
+                let data =
+                    List.init 10 (fun _ ->
+                        {| visible = AVal.init true
+                           trafo   = AVal.init <| randomTrafo() |}
+                    )
+
+                let sg =
+                    AList.ofList data
+                    |> AList.chooseA (fun data ->
+                        data.visible |> AVal.map (fun visible ->
+                            if visible then
+                                Some <| drawLeaf data.trafo positions
+                            else
+                                None
+                        )
+                    )
+                    |> AList.toASet
+                    |> ASet.collect _.RenderObjects(Ag.Scope.Root)
+                    |> Sg.renderObjectSet
+                    |> Sg.trafo globalTrafo
+
+                let expected =
+                    AVal.custom (fun t ->
+                        (Box3d.Invalid, data) ||> List.fold (fun result data ->
+                            if data.visible.GetValue t then
+                                let trafo = data.trafo.GetValue t * globalTrafo.GetValue t
+                                let bb = positions |> Array.map (V3d >> Mat.transformPos trafo.Forward) |> Box3d
+                                Box.Union(result, bb)
+                            else
+                                result
+                        )
+                    )
+
+                let actualLocal = sg.LocalBoundingBox Ag.Scope.Root
+                let actualGlobal = sg.GlobalBoundingBox Ag.Scope.Root
+
+                let randomize() =
+                    transact (fun _ ->
+                        if Rnd.bool() then globalTrafo.Value <- randomTrafo()
+
+                        for d in data do
+                            d.visible.Value <- Rnd.bool()
+                            if Rnd.bool() then d.trafo.Value <- randomTrafo()
+                    )
+
+                let test() =
+                    let expected = expected.GetValue()
+                    let actualLocal = actualLocal.GetValue()
+                    let actualGlobal = actualGlobal.GetValue()
+                    Expect.approxEquals actualLocal expected 0.001 "Invalid local bounding box"
+                    Expect.approxEquals actualGlobal expected 0.001 "Invalid global bounding box"
+
+                test()
+
+                for _ = 1 to 100 do
+                    randomize()
+                    test()
+
+                transact (fun _ -> for d in data do d.visible.Value <- false)
+                test()
+            }
+
+        let private commands<'TCommand> (name: string) (toSg: 'TCommand -> ISg)
+                                        (clearCmd: C4b -> 'TCommand) (ifThenElseCmd: aval<bool> * 'TCommand * 'TCommand -> 'TCommand)
+                                        (orderedCmd: alist<'TCommand> -> 'TCommand) (unorderedCmd: list<ISg> -> 'TCommand) =
+            test $"Bounding Box.{name}" {
+                IntrospectionProperties.CustomEntryAssembly <- Assembly.GetAssembly(typeof<ISg>)
+                Aardvark.Init()
+
+                let positions = [| V3f(-0.5f, -0.25f, 0.0f); V3f(0.0f, -10.0f, 5.0f); V3f(3.0f, 7.0f, -7.0f) |]
+                let globalTrafo = AVal.init <| randomTrafo()
+
+                let data =
+                    List.init 10 (fun _ ->
+                        {| visible = AVal.init true
+                           switch  = AVal.init true
+                           t1      = AVal.init <| randomTrafo()
+                           t2      = AVal.init <| randomTrafo() |}
+                    )
+
+                let sg =
+                    orderedCmd (alist {
+                        yield clearCmd C4b.Blue
+
+                        for d in data do
+                            match! d.visible with
+                            | true ->
+                                let l1 = unorderedCmd [drawLeaf d.t1 positions]
+                                let l2 = unorderedCmd [drawLeaf d.t2 positions]
+                                yield ifThenElseCmd(d.switch, l1, l2)
+
+                            | _ -> ()
+                    })
+                    |> toSg
+                    |> Sg.trafo globalTrafo
+
+                let expected =
+                    AVal.custom (fun t ->
+                        (Box3d.Invalid, data) ||> List.fold (fun result data ->
+                            if data.visible.GetValue t then
+                                let localTrafo = if data.switch.GetValue t then data.t1 else data.t2
+                                let trafo = localTrafo.GetValue t * globalTrafo.GetValue t
+                                let bb = positions |> Array.map (V3d >> Mat.transformPos trafo.Forward) |> Box3d
+                                Box.Union(result, bb)
+                            else
+                                result
+                        )
+                    )
+
+                let actualLocal = sg.LocalBoundingBox Ag.Scope.Root
+                let actualGlobal = sg.GlobalBoundingBox Ag.Scope.Root
+
+                let randomize() =
+                    transact (fun _ ->
+                        if Rnd.bool() then globalTrafo.Value <- randomTrafo()
+
+                        for d in data do
+                            d.visible.Value <- Rnd.bool()
+                            d.switch.Value <- Rnd.bool()
+                            if Rnd.bool() then
+                                d.t1.Value <- randomTrafo()
+                                d.t2.Value <- randomTrafo()
+                    )
+
+                let test() =
+                    let expected = expected.GetValue()
+                    let actualLocal = actualLocal.GetValue()
+                    let actualGlobal = actualGlobal.GetValue()
+                    Expect.approxEquals actualLocal expected 0.001 "Invalid local bounding box"
+                    Expect.approxEquals actualGlobal expected 0.001 "Invalid global bounding box"
+
+                test()
+
+                for _ = 1 to 100 do
+                    randomize()
+                    test()
+
+                transact (fun _ -> for d in data do d.visible.Value <- false)
+                test()
+            }
+
+        let renderCommands =
+            commands "RenderCommands"
+                Sg.execute
+                RenderCommand.Clear
+                RenderCommand.IfThenElse
+                RenderCommand.Ordered
+                RenderCommand.Unordered
+
+        let runtimeCommands =
+            let toSg (cmd: RuntimeCommand) =
+                let ro = CommandRenderObject(RenderPass.main, Ag.Scope.Root, cmd) :> IRenderObject
+                ro |> ASet.single |> Sg.renderObjectSet
+
+            let clearCmd (color: C4b) =
+                let values = ClearValues.ofColor color
+                RuntimeCommand.Clear (AVal.constant values)
+
+            let unorderedCmd (sg: ISg list) =
+                let ro =
+                    sg
+                    |> List.map _.RenderObjects(Ag.Scope.Root)
+                    |> ASet.ofList
+                    |> ASet.unionMany
+
+                RuntimeCommand.Render ro
+
+            commands "RuntimeCommands"
+                toSg
+                clearCmd
+                RuntimeCommand.IfThenElse
+                RuntimeCommand.Ordered
+                unorderedCmd
+
     [<Tests>]
     let tests =
         testList "SceneGraph" [
             checkCompleteness
             onActivationMultiRenderObject
             delayModifySurface
+            modelTrafo
+            BoundingBox.renderNode
+            BoundingBox.renderObjectsNode
+            BoundingBox.renderCommands
+            BoundingBox.runtimeCommands
         ]